Memoize sidebar
