+++ components/sidebar/SidebarNotesFooter.tsx
@@ -1,17 +1,11 @@
-import { useRef, useState, Dispatch, SetStateAction, useCallback } from 'react';
-import { Menu } from '@headlessui/react';
-import {
-  IconSortDescending,
-  IconCheck,
-  IconPlus,
-  IconDownload,
-} from '@tabler/icons';
-import { usePopper } from 'react-popper';
+import { Dispatch, SetStateAction, useCallback, memo } from 'react';
+import { IconPlus, IconDownload } from '@tabler/icons';
 import { useStore } from 'lib/store';
 import useImport from 'utils/useImport';
-import { ReadableNameBySort, Sort } from 'lib/createUserSettingsSlice';
+import { Sort } from 'lib/createUserSettingsSlice';
 import Tooltip from 'components/Tooltip';
 import { isMobile } from 'utils/device';
+import SidebarNotesSortDropdown from './SidebarNotesSortDropdown';
 
 type Props = {
   noteSort: Sort;
@@ -19,7 +13,7 @@ type Props = {
   setIsFindOrCreateModalOpen: Dispatch<SetStateAction<boolean>>;
 };
 
-export default function SidebarNotesFooter(props: Props) {
+function SidebarNotesFooter(props: Props) {
   const { noteSort, numOfNotes, setIsFindOrCreateModalOpen } = props;
   const onImport = useImport();
 
@@ -46,7 +40,10 @@ export default function SidebarNotesFooter(props: Props) {
         {numOfNotes} notes
       </span>
       <div className="mx-2 my-1">
-        <SortDropdown currentSort={noteSort} setCurrentSort={setNoteSort} />
+        <SidebarNotesSortDropdown
+          currentSort={noteSort}
+          setCurrentSort={setNoteSort}
+        />
         <Tooltip content="Import">
           <button
             className="p-1 rounded hover:bg-gray-200 active:bg-gray-300 dark:hover:bg-gray-700 dark:active:bg-gray-600"
@@ -63,74 +60,4 @@ export default function SidebarNotesFooter(props: Props) {
   );
 }
 
-type SortDropdownProps = {
-  currentSort: Sort;
-  setCurrentSort: (sort: Sort) => void;
-};
-
-const SortDropdown = (props: SortDropdownProps) => {
-  const { currentSort, setCurrentSort } = props;
-
-  const buttonRef = useRef<HTMLButtonElement | null>(null);
-  const [popperElement, setPopperElement] = useState<HTMLDivElement | null>(
-    null
-  );
-  const { styles, attributes } = usePopper(buttonRef.current, popperElement, {
-    placement: 'bottom-start',
-  });
-
-  return (
-    <Menu>
-      <Tooltip content="Sort notes">
-        <Menu.Button
-          className="p-1 rounded hover:bg-gray-200 active:bg-gray-300 dark:hover:bg-gray-700 dark:active:bg-gray-600"
-          ref={buttonRef}
-        >
-          <IconSortDescending
-            size={16}
-            className="text-gray-600 dark:text-gray-300"
-          />
-        </Menu.Button>
-      </Tooltip>
-      <Menu.Items
-        className="z-10 w-56 overflow-hidden bg-white rounded dark:bg-gray-800 shadow-popover"
-        ref={setPopperElement}
-        style={styles.popper}
-        {...attributes.popper}
-      >
-        {Object.values(Sort).map((sort, index, arr) => {
-          const isActive = currentSort === sort;
-          const showDivider = (index + 1) % 2 === 0 && index !== arr.length - 1;
-          return (
-            <Menu.Item key={sort}>
-              {({ active }) => (
-                <button
-                  className={`flex w-full items-center px-4 py-2 text-left text-gray-800 dark:text-gray-200 text-sm ${
-                    showDivider ? 'border-b dark:border-gray-700' : ''
-                  } ${active ? 'bg-gray-100 dark:bg-gray-700' : ''}`}
-                  onClick={() => setCurrentSort(sort)}
-                >
-                  <span
-                    className={
-                      isActive
-                        ? 'text-primary-600 dark:text-primary-400'
-                        : undefined
-                    }
-                  >
-                    {ReadableNameBySort[sort]}
-                  </span>
-                  {isActive ? (
-                    <IconCheck
-                      size={18}
-                      className="ml-1 text-primary-600 dark:text-primary-400"
-                    />
-                  ) : null}
-                </button>
-              )}
-            </Menu.Item>
-          );
-        })}
-      </Menu.Items>
-    </Menu>
-  );
-};
+export default memo(SidebarNotesFooter);
